github api client: hoist records out of GitHubApiClient
MintedToken / InstallationDetails / CreatedRepository become top-level
records in the same package. Easier to find, easier to import, no nested
class noise in the api client. No behaviour change.
