+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/GitHubProjectRepoProvisioner.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.kinotic.github.api.config.KinoticGithubProperties;
 import org.kinotic.github.api.model.GitHubAppInstallation;
 import org.kinotic.github.api.services.GitHubAppInstallationService;
+import org.kinotic.github.internal.api.services.client.CreatedRepository;
 import org.kinotic.github.internal.api.services.client.GitHubApiClient;
-import org.kinotic.github.internal.api.services.client.GitHubApiClient.CreatedRepository;
 import org.kinotic.github.internal.api.services.client.GitHubInstallationTokenCache;
 import org.kinotic.os.api.model.Project;
 import org.kinotic.os.api.model.RepositoryConnectionStatus;
