+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/client/CreatedRepository.java
@@ -1,0 +1,8 @@
+package org.kinotic.github.internal.api.services.client;
+
+/**
+ * Subset of GitHub's repository JSON returned by
+ * {@link GitHubApiClient#createRepoFromTemplate}. Carries the fields the
+ * provisioner stamps onto the new {@code Project}.
+ */
+public record CreatedRepository(Long id, String fullName, String defaultBranch) {}
