+++ kinotic-github/src/main/java/org/kinotic/github/internal/api/services/client/InstallationDetails.java
@@ -1,0 +1,8 @@
+package org.kinotic.github.internal.api.services.client;
+
+/**
+ * Subset of GitHub's installation JSON the platform reads — the install id plus
+ * the owning account's login + type ({@code User} or {@code Organization}).
+ * Returned by {@link GitHubApiClient#getInstallation}.
+ */
+public record InstallationDetails(Long id, String accountLogin, String accountType) {}
